refactor: split recipe-run lifecycle helpers
