+++ packages/cli/src/commands/recipe-run.ts
@@ -1,233 +1,26 @@
 import { createHash } from "node:crypto"
-import { fstatSync } from "node:fs"
-import { mkdir, readdir, readFile, stat, writeFile } from "node:fs/promises"
-import { basename, dirname, join, relative, resolve } from "node:path"
-import { setTimeout as delay } from "node:timers/promises"
-import { DEFAULT_WORDPRESS_VERSION, RuntimeRunRegistry, artifactBundleRunRef, artifactManifestFile, createBenchResultsJsonSchema, createRuntimeRunId, defaultRunRegistryDirectory, createRuntime, refreshArtifactManifestFileSha256s, stripUndefined, upsertArtifactManifestFiles, type ArtifactBundle, type ArtifactManifest, type ArtifactManifestFile, type BenchmarkArtifactRef, type BenchResults, type ExecutionResult, type Runtime, type RuntimeAssetSpec, type RuntimeInfo, type RuntimePreviewSpec, type RuntimeRunRecord, type WorkspaceRecipe, type WorkspaceRecipeDeclaredArtifact, type WorkspaceRecipeFixtureDatabase, type WorkspaceRecipePluginRuntimeHealthProbe, type WorkspaceRecipeProbe, type WorkspaceRecipeSiteSeed } from "@automattic/wp-codebox-core"
+import { readdir, readFile, stat, writeFile } from "node:fs/promises"
+import { basename, dirname, join, resolve } from "node:path"
+import { DEFAULT_WORDPRESS_VERSION, artifactBundleRunRef, artifactManifestFile, createBenchResultsJsonSchema, createRuntime, refreshArtifactManifestFileSha256s, stripUndefined, upsertArtifactManifestFiles, type ArtifactBundle, type ArtifactManifest, type ArtifactManifestFile, type BenchmarkArtifactRef, type BenchResults, type ExecutionResult, type Runtime, type RuntimeAssetSpec, type RuntimePreviewSpec, type RuntimeRunRecord, type RuntimeRunRegistry, type WorkspaceRecipe, type WorkspaceRecipeDeclaredArtifact, type WorkspaceRecipeFixtureDatabase, type WorkspaceRecipePluginRuntimeHealthProbe, type WorkspaceRecipeProbe, type WorkspaceRecipeSiteSeed } from "@automattic/wp-codebox-core"
 import { createPlaygroundRuntimeBackend } from "@automattic/wp-codebox-playground"
 import { Ajv2020 } from "ajv/dist/2020.js"
 import { recipeExecutionSpec, sandboxWorkspaceContract } from "../agent-sandbox.js"
 import { executeAgentFanoutFromArgs } from "../agent-fanout.js"
 import { captureStdout, printRecipeHumanOutput, printRecipeValidateHumanOutput, serializeError } from "../output.js"
 import { parsePreviewBind, parsePreviewHoldSeconds, parsePreviewPort, parsePreviewPublicUrl } from "../preview-options.js"
-import { dryRunRecipe, planWorkspaceRecipe, pluginRuntimeHealthProbeStepIndex, pluginRuntimeSetupStepIndex, recipeDryRunSiteSeeds, siteSeedScopesAreBounded, type RecipeDryRunOutput, type RecipeDryRunSiteSeed, type RecipeDryRunStagedFile } from "../recipe-dry-run.js"
-import { appendRecipeRuntimeEvidence, collectAndFinalizeFailedRecipeArtifacts, finalizeAgentSandboxEvidence, finalizeRecipeArtifactEvidence, recipeAgentResultFailure, recipeAgentResultOutput, recipeAgentTaskResultOutput, recipeArtifactEvidenceFailure, recipeCompletionOutcomeOutput, type AgentSandboxResultSummary, type AgentTaskSingleResult, type SandboxCompletionOutcome } from "../recipe-evidence.js"
+import { dryRunRecipe, planWorkspaceRecipe, pluginRuntimeHealthProbeStepIndex, pluginRuntimeSetupStepIndex, recipeDryRunSiteSeeds, siteSeedScopesAreBounded } from "../recipe-dry-run.js"
+import { appendRecipeRuntimeEvidence, collectAndFinalizeFailedRecipeArtifacts, finalizeAgentSandboxEvidence, finalizeRecipeArtifactEvidence, recipeAgentResultFailure, recipeAgentResultOutput, recipeAgentTaskResultOutput, recipeArtifactEvidenceFailure, recipeCompletionOutcomeOutput } from "../recipe-evidence.js"
 import { prepareRecipeRuntimeBackendPackage, type PreparedRuntimeBackendPackage } from "../recipe-backend-package.js"
 import { cleanupRecipePreparedSources, installMuPluginsCode, prepareRecipeExtraPlugins, prepareRecipeRuntimeOverlays, prepareRecipeStagedFiles, prepareRecipeWorkspaces, recipeBlueprintWithBootActivePlugins, recipeExtraPlugins, recipeMountType, type PreparedExtraPlugin, type PreparedRuntimeOverlay, type PreparedStagedFile, type PreparedWorkspaceMount } from "../recipe-sources.js"
-import { loadWorkspaceRecipe, pluginRuntimeHealthProbeStep, recipePolicy, recipeWorkflowSteps, validateWorkspaceRecipe, type RecipeValidationIssue, type RecipeWorkflowPhase } from "../recipe-validation.js"
+import { loadWorkspaceRecipe, pluginRuntimeHealthProbeStep, recipePolicy, recipeWorkflowSteps, validateWorkspaceRecipe, type RecipeWorkflowPhase } from "../recipe-validation.js"
 import { previewSpec, releaseRuntime, runtimeMetadata, type RunOutput } from "../runtime-command-wrappers.js"
-
-interface RecipeRunOptions {
-  recipePath: string
-  artifactsDirectory?: string
-  runRegistryDirectory?: string
-  previewHoldSeconds?: number
-  previewPublicUrl?: string
-  previewPort?: number
-  previewBind?: string
-  timeoutMs: number
-  json: boolean
-  dryRun: boolean
-}
-
-type RecipeInterruptionSignal = "SIGINT" | "SIGTERM" | "SIGHUP"
-type RecipeInterruptionReason = "signal" | "parent-disconnect" | "stdio-closed"
-
-interface RecipeInterruptionMetadata {
-  signal: RecipeInterruptionSignal
-  reason: RecipeInterruptionReason
-  receivedAt: string
-  artifactsFinalized: boolean
-}
-
-interface RecipeInterruptionController {
-  readonly metadata: RecipeInterruptionMetadata | undefined
-  install(): void
-  dispose(): void
-  interruptible<T>(promise: Promise<T>): Promise<T>
-  throwIfInterrupted(): void
-  propagateIfInterrupted(): void
-}
-
-interface RecipeValidateOptions {
-  recipePath: string
-  json: boolean
-}
-
-interface RecipeValidateOutput {
-  success: boolean
-  schema: "wp-codebox/recipe-validation/v1"
-  recipePath?: string
-  valid: boolean
-  issues: RecipeValidationIssue[]
-  summary?: {
-    steps: number
-    mounts: number
-    workspaces: number
-    extraPlugins: number
-    stagedFiles: number
-  }
-  error?: RunOutput["error"]
-}
-
-interface RecipeRunOutput {
-  success: boolean
-  schema: "wp-codebox/recipe-run/v1"
-  recipePath?: string
-  runtime?: RuntimeInfo
-  executions: RecipeExecutionResult[]
-  stagedFiles?: RecipeRunStagedFile[]
-  fixtureDatabases?: RecipeRunFixtureDatabase[]
-  siteSeeds?: RecipeRunSiteSeed[]
-  probes?: RecipeRunProbe[]
-  declaredArtifacts?: RecipeRunDeclaredArtifact[]
-  phaseEvidence?: RecipePhaseEvidence[]
-  diagnostics?: RecipeRuntimeDiagnostic[]
-  validation?: {
-    issues: RecipeValidationIssue[]
-  }
-  benchResults?: BenchResults
-  benchResultsList?: BenchResults[]
-  agentResult?: AgentSandboxResultSummary
-  agentTaskResult?: AgentTaskSingleResult
-  completionOutcome?: SandboxCompletionOutcome
-  artifacts?: ArtifactBundle
-  run?: RuntimeRunRecord
-  interruption?: RecipeInterruptionMetadata
-  logs?: string[]
-  error?: RunOutput["error"]
-}
-
-type RecipeRunCommandOutput = RecipeRunOutput | RecipeDryRunOutput
+import { createRecipeRunContext } from "./recipe-run-context.js"
+import { createRecipeInterruptionController, interruptedRecipeOutput, markRecipeArtifactsFinalized, recipeInterruptionSerializedError } from "./recipe-run-interruption.js"
+import { bestEffortTimeout, exitAfterPlaygroundCliBootFailure, exitAfterRecipeRunTimeout, isRecipeRunTimeoutError, printJsonFailureDiagnostic, recipeRunFailureStatus, RecipeDeclaredArtifactFailureError, RecipeProbeFailureError, RecipeRunTimeoutError, RecipeRuntimeCreateError, remainingRecipeTimeoutMs, serializeRecipeRunError, watchRecipeOperation } from "./recipe-run-output.js"
+import { RecipePhaseError, RecipePhaseTracker } from "./recipe-run-phases.js"
+import type { RecipeExecutionResult, RecipeInterruptionController, RecipePhaseEvidence, RecipePhaseName, RecipePhpWasmRuntimeDiagnostic, RecipeRunCommandOutput, RecipeRunDeclaredArtifact, RecipeRunFixtureDatabase, RecipeRunOptions, RecipeRunOutput, RecipeRunProbe, RecipeRunSiteSeed, RecipeRunStagedFile, RecipeRuntimeDiagnostic, RecipeValidateOptions, RecipeValidateOutput } from "./recipe-run-types.js"
 
 const DEFAULT_RECIPE_RUN_TIMEOUT_MS = 25 * 60 * 1000
-
-type RecipeExecutionResult = ExecutionResult & {
-  recipePhase?: RecipeWorkflowPhase
-  recipeStepIndex?: number
-  recipeCommand?: string
-}
-
-type RecipeArtifactPointerCommandStatus = "queued" | "running" | "completed" | "failed"
-
-interface RecipeArtifactPointerState {
-  command?: string
-  commandStatus?: RecipeArtifactPointerCommandStatus
-  runtime?: RuntimeInfo
-  artifacts?: ArtifactBundle
-  failure?: RunOutput["error"]
-  phases?: RecipePhaseEvidence[]
-}
-
-type RecipePhaseName = "runtime_startup" | "mount_plugins" | "activate_plugins" | "run_blueprint_steps" | "import_fixture_databases" | "run_workloads" | "run_probes" | "collect_artifacts"
-
-interface RecipePhaseEvidence {
-  schema: "wp-codebox/recipe-phase-evidence/v1"
-  name: RecipePhaseName
-  status: "completed" | "failed"
-  startedAt: string
-  endedAt: string
-  durationMs: number
-  data?: Record<string, unknown>
-  error?: RunOutput["error"]
-}
-
-interface RecipePluginRuntimeDiagnostic {
-  schema: "wp-codebox/plugin-runtime-diagnostic/v1"
-  severity: "error"
-  phase: "setup" | "health-probe" | "runtime" | "overlay-preparation" | "backend-preparation"
-  name?: string
-  command?: string
-  exitCode?: number
-  message: string
-  executionIndex?: number
-}
-
-interface RecipePhpWasmRuntimeDiagnostic {
-  schema: "wp-codebox/php-wasm-runtime-diagnostic/v1"
-  severity: "error"
-  phase: "preflight"
-  message: string
-  runtime?: Record<string, unknown>
-  repair?: string
-}
-
-interface RecipePhaseDiagnostic {
-  schema: "wp-codebox/recipe-phase-diagnostic/v1"
-  severity: "error"
-  phase: RecipePhaseName
-  pluginFile?: string
-  command?: string
-  exitCode?: number
-  message: string
-  executionIndex?: number
-}
-
-type RecipeRuntimeDiagnostic = RecipePluginRuntimeDiagnostic | RecipePhaseDiagnostic | RecipePhpWasmRuntimeDiagnostic
-
-interface RecipeRunSiteSeed extends Omit<RecipeDryRunSiteSeed, "dryRunOnly"> {
-  action: "imported" | "skipped"
-  reason?: string
-  counts?: Record<string, number>
-  warnings?: string[]
-  provenance?: Record<string, unknown>
-}
-
-interface RecipeRunStagedFile extends RecipeDryRunStagedFile {
-  action: "staged"
-}
-
-interface RecipeRunFixtureDatabase {
-  schema: "wp-codebox/fixture-database-result/v1"
-  index: number
-  name: string
-  version: string
-  source: string
-  format: "sql"
-  action: "imported" | "skipped"
-  reset: {
-    strategy: "none" | "truncate-tables"
-    tables: string[]
-  }
-  identity: {
-    name: string
-    version: string
-    sourceSha256: string
-  }
-  counts?: Record<string, number>
-  metadata?: Record<string, unknown>
-}
-
-interface RecipeRunProbe {
-  schema: "wp-codebox/recipe-probe-result/v1"
-  index: number
-  name: string
-  status: "passed" | "failed"
-  command: string
-  args: string[]
-  exitCode: number
-  stdout: string
-  stderr: string
-  parsedJson?: unknown
-  allowFailure: boolean
-  metadata?: Record<string, unknown>
-}
-
-interface RecipeRunDeclaredArtifact {
-  schema: "wp-codebox/recipe-declared-artifact-result/v1"
-  index: number
-  name: string
-  path: string
-  required: boolean
-  status: "collected" | "missing" | "failed"
-  exists: boolean
-  type?: "file" | "directory" | "other"
-  size?: number
-  sha256?: string
-  parsedJson?: unknown
-  error?: RunOutput["error"]
-  metadata?: Record<string, unknown>
-}
 
 interface BenchmarkArtifactOutput {
   schema: "wp-codebox/benchmark-artifacts/v1"
@@ -283,35 +76,6 @@ export async function runRecipeRunCommand(args: string[]): Promise<number> {
   }
 }
 
-function exitAfterPlaygroundCliBootFailure(output: RecipeRunCommandOutput): void {
-  if (output.schema === "wp-codebox/recipe-run/v1" && hasSerializedErrorCode(output.error, "wp-codebox-playground-cli-exited")) {
-    process.exit(output.success ? 0 : 1)
-  }
-}
-
-function hasSerializedErrorCode(error: RunOutput["error"] | undefined, code: string): boolean {
-  if (!error) {
-    return false
-  }
-
-  if (error.code === code) {
-    return true
-  }
-
-  const cause = error.cause
-  if (!cause || typeof cause !== "object" || Array.isArray(cause)) {
-    return false
-  }
-
-  return hasSerializedErrorCode(cause as RunOutput["error"], code)
-}
-
-function exitAfterRecipeRunTimeout(output: RecipeRunCommandOutput): void {
-  if (output.schema === "wp-codebox/recipe-run/v1" && output.error?.code === "recipe-run-timeout") {
-    process.exit(output.success ? 0 : 1)
-  }
-}
-
 export async function runRecipeValidateCommand(args: string[]): Promise<number> {
   const options = parseRecipeValidateOptions(args)
   const output = await validateRecipe(options)
@@ -324,487 +88,10 @@ export async function runRecipeValidateCommand(args: string[]): Promise<number> 
   return output.success ? 0 : 1
 }
 
-function printJsonFailureDiagnostic(output: { success: boolean; error?: { message?: string }; logs?: string[] }): void {
-  if (output.success) {
-    return
-  }
-
-  const message = output.error?.message?.trim()
-  if (message) {
-    console.error(message)
-  }
-
-  for (const log of output.logs ?? []) {
-    const trimmed = log.trim()
-    if (trimmed) {
-      console.error(trimmed)
-    }
-  }
-}
-
-class RecipeInterruptedError extends Error {
-  readonly code = "recipe-interrupted"
-
-  constructor(readonly signal: RecipeInterruptionSignal, readonly reason: RecipeInterruptionReason, readonly receivedAt: string) {
-    super(recipeInterruptionMessage({ signal, reason }))
-    this.name = "RecipeInterruptedError"
-  }
-}
-
-class RecipeRunTimeoutError extends Error {
-  readonly code = "recipe-run-timeout"
-  readonly activeOperation: string
-  readonly elapsedMs: number
-  readonly timeoutMs: number
-
-  constructor(activeOperation: string, elapsedMs: number, timeoutMs: number) {
-    super(`Recipe run timed out after ${elapsedMs}ms while waiting for ${activeOperation}`)
-    this.name = "RecipeRunTimeoutError"
-    this.activeOperation = activeOperation
-    this.elapsedMs = elapsedMs
-    this.timeoutMs = timeoutMs
-  }
-}
-
-class RecipeRuntimeCreateError extends Error {
-  readonly code = "recipe-runtime-create-failed"
-
-  constructor(message: string, readonly context: Record<string, unknown>, cause: unknown) {
-    super(message, { cause })
-    this.name = "RecipeRuntimeCreateError"
-  }
-}
-
-class RecipePhaseError extends Error {
-  readonly code = "recipe-phase-failed"
-
-  constructor(readonly phase: RecipePhaseName, readonly phaseData: Record<string, unknown> | undefined, cause: unknown) {
-    const message = cause instanceof Error ? cause.message : String(cause)
-    super(`Recipe phase ${phase} failed: ${message}`, { cause })
-    this.name = "RecipePhaseError"
-  }
-}
-
-class RecipeProbeFailureError extends Error {
-  readonly code = "recipe-probe-failed"
-
-  constructor(readonly probes: RecipeRunProbe[]) {
-    const failed = probes.filter((probe) => probe.status === "failed" && !probe.allowFailure)
-    super(`${failed.length} recipe probe${failed.length === 1 ? "" : "s"} failed: ${failed.map((probe) => probe.name).join(", ")}`)
-    this.name = "RecipeProbeFailureError"
-  }
-}
-
-class RecipeDeclaredArtifactFailureError extends Error {
-  readonly code = "recipe-artifact-collection-failed"
-
-  constructor(readonly declaredArtifacts: RecipeRunDeclaredArtifact[]) {
-    const failed = declaredArtifacts.filter((artifact) => artifact.required && artifact.status !== "collected")
-    super(`${failed.length} required recipe artifact${failed.length === 1 ? "" : "s"} failed collection: ${failed.map((artifact) => artifact.name).join(", ")}`)
-    this.name = "RecipeDeclaredArtifactFailureError"
-  }
-}
-
-class RecipePhaseTracker {
-  private phases: RecipePhaseEvidence[] = []
-
-  list(): RecipePhaseEvidence[] {
-    return this.phases
-  }
-
-  complete(name: RecipePhaseName, data?: Record<string, unknown>): void {
-    const now = new Date().toISOString()
-    this.phases.push({
-      schema: "wp-codebox/recipe-phase-evidence/v1",
-      name,
-      status: "completed",
-      startedAt: now,
-      endedAt: now,
-      durationMs: 0,
-      ...(data ? { data } : {}),
-    })
-  }
-
-  fail(name: RecipePhaseName, error: unknown, data?: Record<string, unknown>): void {
-    const now = new Date().toISOString()
-    this.phases.push({
-      schema: "wp-codebox/recipe-phase-evidence/v1",
-      name,
-      status: "failed",
-      startedAt: now,
-      endedAt: now,
-      durationMs: 0,
-      ...(data ? { data } : {}),
-      error: serializeRecipeRunError(error),
-    })
-  }
-
-  async run<T>(name: RecipePhaseName, data: Record<string, unknown> | undefined, callback: () => Promise<T>): Promise<T> {
-    const startedAtMs = Date.now()
-    const startedAt = new Date().toISOString()
-    try {
-      const result = await callback()
-      this.phases.push({
-        schema: "wp-codebox/recipe-phase-evidence/v1",
-        name,
-        status: "completed",
-        startedAt,
-        endedAt: new Date().toISOString(),
-        durationMs: Date.now() - startedAtMs,
-        ...(data ? { data } : {}),
-      })
-      return result
-    } catch (error) {
-      const phaseError = error instanceof RecipePhaseError || error instanceof RecipeRunTimeoutError ? error : new RecipePhaseError(name, data, error)
-      this.phases.push({
-        schema: "wp-codebox/recipe-phase-evidence/v1",
-        name,
-        status: "failed",
-        startedAt,
-        endedAt: new Date().toISOString(),
-        durationMs: Date.now() - startedAtMs,
-        ...(data ? { data } : {}),
-        error: serializeRecipeRunError(error),
-      })
-      throw phaseError
-    }
-  }
-}
-
-class RecipeArtifactPointerTracker {
-  private command: string | undefined
-  private commandStatus: RecipeArtifactPointerCommandStatus = "queued"
-  private runtime: RuntimeInfo | undefined
-  private artifacts: ArtifactBundle | undefined
-  private failure: RunOutput["error"] | undefined
-  private phases: RecipePhaseEvidence[] = []
-
-  constructor(private readonly directory: string | undefined, private readonly runId: string, private readonly recipePath: string, private readonly startedAt: string) {}
-
-  async update(state: RecipeArtifactPointerState = {}): Promise<void> {
-    if (!this.directory) {
-      return
-    }
-
-    this.command = state.command ?? this.command
-    this.commandStatus = state.commandStatus ?? this.commandStatus
-    this.runtime = state.runtime ?? this.runtime
-    this.artifacts = state.artifacts ?? this.artifacts
-    this.failure = state.failure ?? this.failure
-    this.phases = state.phases ?? this.phases
-
-    const pointer = stripUndefined({
-      schema: "wp-codebox/recipe-run-artifact-pointer/v1",
-      runId: this.runId,
-      recipePath: this.recipePath,
-      startedAt: this.startedAt,
-      updatedAt: new Date().toISOString(),
-      runtimeId: this.runtime?.id,
-      runtime: this.runtime,
-      currentCommand: this.commandStatus === "running" ? this.command : undefined,
-      lastCommand: this.command,
-      commandStatus: this.commandStatus,
-      failure: this.failure,
-      failurePhase: recipeArtifactPointerFailurePhase(this.failure, this.phases),
-      paths: recipeArtifactPointerPaths(this.directory, this.runtime, this.artifacts),
-    })
-
-    await mkdir(this.directory, { recursive: true })
-    const contents = `${JSON.stringify(pointer, null, 2)}\n`
-    await writeFile(join(this.directory, "latest-runtime.json"), contents)
-    await writeFile(join(this.directory, "manifest.json"), contents)
-  }
-}
-
-function createRecipeInterruptionController(): RecipeInterruptionController {
-  let metadata: RecipeInterruptionMetadata | undefined
-  let rejectInterrupted: ((error: RecipeInterruptedError) => void) | undefined
-  let installed = false
-  let parentWatcher: NodeJS.Timeout | undefined
-  let stdinWatcherInstalled = false
-  let stdioErrorWatcherInstalled = false
-  const initialParentPid = process.ppid
-  const signals: RecipeInterruptionSignal[] = ["SIGINT", "SIGTERM", "SIGHUP"]
-  const interrupt = (signal: RecipeInterruptionSignal, reason: RecipeInterruptionReason): void => {
-    if (!metadata) {
-      metadata = { signal, reason, receivedAt: new Date().toISOString(), artifactsFinalized: false }
-    }
-    rejectInterrupted?.(new RecipeInterruptedError(metadata.signal, metadata.reason, metadata.receivedAt))
-  }
-  const handler = (signal: RecipeInterruptionSignal): void => {
-    interrupt(signal, "signal")
-  }
-  const parentDisconnectHandler = (): void => {
-    interrupt("SIGHUP", "parent-disconnect")
-  }
-  const stdinClosedHandler = (): void => {
-    interrupt("SIGHUP", "stdio-closed")
-  }
-  const stdioErrorHandler = (error: NodeJS.ErrnoException): void => {
-    if (error.code === "EPIPE" || error.code === "ERR_STREAM_DESTROYED") {
-      interrupt("SIGHUP", "stdio-closed")
-    }
-  }
-
-  const controller: RecipeInterruptionController = {
-    get metadata() {
-      return metadata
-    },
-    install() {
-      if (installed) {
-        return
-      }
-      for (const signal of signals) {
-        process.on(signal, handler)
-      }
-      if (initialParentPid > 1) {
-        parentWatcher = setInterval(() => {
-          if (process.ppid === 1 || process.ppid !== initialParentPid) {
-            parentDisconnectHandler()
-          }
-        }, 1_000)
-        parentWatcher.unref()
-      }
-      if (!process.stdin.isTTY && process.stdin.readable && stdinCanSignalParentDisconnect()) {
-        process.stdin.on("end", stdinClosedHandler)
-        process.stdin.on("close", stdinClosedHandler)
-        process.stdin.resume()
-        stdinWatcherInstalled = true
-      }
-      process.stdout.on("error", stdioErrorHandler)
-      process.stderr.on("error", stdioErrorHandler)
-      stdioErrorWatcherInstalled = true
-      installed = true
-    },
-    dispose() {
-      if (!installed) {
-        return
-      }
-      for (const signal of signals) {
-        process.off(signal, handler)
-      }
-      if (parentWatcher) {
-        clearInterval(parentWatcher)
-        parentWatcher = undefined
-      }
-      if (stdinWatcherInstalled) {
-        process.stdin.off("end", stdinClosedHandler)
-        process.stdin.off("close", stdinClosedHandler)
-        process.stdin.pause()
-        stdinWatcherInstalled = false
-      }
-      if (stdioErrorWatcherInstalled) {
-        process.stdout.off("error", stdioErrorHandler)
-        process.stderr.off("error", stdioErrorHandler)
-        stdioErrorWatcherInstalled = false
-      }
-      installed = false
-    },
-    async interruptible<T>(promise: Promise<T>): Promise<T> {
-      if (metadata) {
-        throw new RecipeInterruptedError(metadata.signal, metadata.reason, metadata.receivedAt)
-      }
-
-      let settled = false
-      try {
-        return await Promise.race([
-          promise.finally(() => {
-            settled = true
-          }),
-          new Promise<T>((_resolve, reject) => {
-            rejectInterrupted = (error) => {
-              if (!settled) {
-                reject(error)
-              }
-            }
-          }),
-        ])
-      } finally {
-        rejectInterrupted = undefined
-      }
-    },
-    throwIfInterrupted() {
-      if (metadata) {
-        throw new RecipeInterruptedError(metadata.signal, metadata.reason, metadata.receivedAt)
-      }
-    },
-    propagateIfInterrupted() {
-      if (!metadata || metadata.reason !== "signal") {
-        return
-      }
-      controller.dispose()
-      process.kill(process.pid, metadata.signal)
-    },
-  }
-
-  return controller
-}
-
-function stdinCanSignalParentDisconnect(): boolean {
-  if (process.env.WP_CODEBOX_RECIPE_RUN_STDIN_DISCONNECT !== "1") {
-    return false
-  }
-
-  try {
-    const stats = fstatSync(0)
-    return stats.isFIFO() || stats.isSocket()
-  } catch {
-    return false
-  }
-}
-
-function markRecipeArtifactsFinalized(interruption: RecipeInterruptionController | undefined, artifactsFinalized: boolean): void {
-  if (interruption?.metadata) {
-    interruption.metadata.artifactsFinalized = artifactsFinalized
-  }
-}
-
-function interruptedRecipeOutput<T extends RecipeRunCommandOutput>(output: T, interruption: RecipeInterruptionController | undefined): T {
-  if (!interruption?.metadata || output.schema !== "wp-codebox/recipe-run/v1") {
-    return output
-  }
-
-  return {
-    ...output,
-    success: false,
-    interruption: interruption.metadata,
-    error: recipeInterruptionSerializedError(interruption.metadata),
-  } as T
-}
-
-function recipeInterruptionSerializedError(metadata: RecipeInterruptionMetadata): RunOutput["error"] {
-  return {
-    name: "RecipeInterruptedError",
-    message: recipeInterruptionMessage(metadata),
-    code: "recipe-interrupted",
-  }
-}
-
-function recipeInterruptionMessage(metadata: Pick<RecipeInterruptionMetadata, "signal" | "reason">): string {
-  if (metadata.reason === "parent-disconnect") {
-    return "Recipe run interrupted after parent process disconnected"
-  }
-  if (metadata.reason === "stdio-closed") {
-    return "Recipe run interrupted after stdio closed"
-  }
-  return `Recipe run interrupted by ${metadata.signal}`
-}
-
-function remainingRecipeTimeoutMs(startedAtMs: number, timeoutMs: number): number {
-  return Math.max(1, timeoutMs - (Date.now() - startedAtMs))
-}
-
-async function watchRecipeOperation<T>(operation: string, promise: Promise<T>, startedAtMs: number, timeoutMs: number, configuredTimeoutMs = timeoutMs): Promise<T> {
-  let timeout: NodeJS.Timeout | undefined
-  promise.catch(() => undefined)
-
-  try {
-    return await Promise.race([
-      promise,
-      new Promise<T>((_resolve, reject) => {
-        timeout = setTimeout(() => {
-          reject(new RecipeRunTimeoutError(operation, Date.now() - startedAtMs, configuredTimeoutMs))
-        }, timeoutMs)
-        timeout.unref()
-      }),
-    ])
-  } finally {
-    if (timeout) {
-      clearTimeout(timeout)
-    }
-  }
-}
-
-async function bestEffortTimeout<T>(promise: Promise<T>, timeoutMs: number): Promise<T | undefined> {
-  promise.catch(() => undefined)
-  return Promise.race([
-    promise,
-    delay(timeoutMs).then(() => undefined),
-  ])
-}
-
-function serializeRecipeRunError(error: unknown): RunOutput["error"] {
-  if (error instanceof RecipePhaseError && (error.cause instanceof RecipeProbeFailureError || error.cause instanceof RecipeDeclaredArtifactFailureError)) {
-    return {
-      ...serializeError(error),
-      code: error.cause.code,
-      cause: serializeError(error.cause),
-    }
-  }
-
-  const serialized = serializeError(error)
-  if (error instanceof RecipeRunTimeoutError) {
-    return {
-      ...serialized,
-      activeOperation: error.activeOperation,
-      elapsedMs: error.elapsedMs,
-      timeoutMs: error.timeoutMs,
-    }
-  }
-
-  return serialized
-}
-
-function recipeArtifactPointerFailurePhase(error: RunOutput["error"] | undefined, phases: RecipePhaseEvidence[]): string | undefined {
-  const failedPhase = [...phases].reverse().find((phase) => phase.status === "failed")
-  if (failedPhase) {
-    return failedPhase.name
-  }
-
-  if (typeof error?.activeOperation === "string") {
-    return error.activeOperation
-  }
-
-  return undefined
-}
-
-function recipeArtifactPointerPaths(directory: string, runtime: RuntimeInfo | undefined, artifacts: ArtifactBundle | undefined): Record<string, string> {
-  const paths: Record<string, string | undefined> = {}
-  if (runtime) {
-    const runtimeDirectory = join(directory, runtime.id)
-    paths.runtimeDirectory = relative(directory, runtimeDirectory) || "."
-    paths.eventLog = relative(directory, join(runtimeDirectory, "events.jsonl"))
-    paths.commandLog = relative(directory, join(runtimeDirectory, "logs", "commands.log"))
-    paths.runtimeLog = relative(directory, join(runtimeDirectory, "logs", "runtime.log"))
-    paths.runtimeMetadata = relative(directory, join(runtimeDirectory, "metadata.json"))
-    paths.runtimeManifest = relative(directory, join(runtimeDirectory, "manifest.json"))
-    paths.browserArtifacts = relative(directory, join(runtimeDirectory, "files", "browser"))
-  }
-
-  if (artifacts) {
-    paths.runtimeDirectory = relative(directory, artifacts.directory) || "."
-    paths.eventLog = relative(directory, artifacts.eventsPath)
-    paths.commandLog = relative(directory, artifacts.commandsLogPath)
-    paths.runtimeLog = relative(directory, artifacts.runtimeLogPath)
-    paths.runtimeMetadata = relative(directory, artifacts.metadataPath)
-    paths.runtimeManifest = relative(directory, artifacts.manifestPath)
-    paths.browserArtifacts = relative(directory, join(artifacts.directory, "files", "browser"))
-  }
-
-  return stripUndefined(paths) as Record<string, string>
-}
-
 async function runRecipe(options: RecipeRunOptions, interruption?: RecipeInterruptionController): Promise<RecipeRunOutput> {
-  const recipePath = resolve(options.recipePath)
-  const recipeDirectory = dirname(recipePath)
-  const recipe = await loadWorkspaceRecipe(recipePath)
-  const configuredArtifactsDirectory = options.artifactsDirectory ?? recipe.artifacts?.directory
-  const runRegistry = new RuntimeRunRegistry(options.runRegistryDirectory ?? defaultRunRegistryDirectory(configuredArtifactsDirectory))
-  const startedAtMs = Date.now()
-  let runRecord = await runRegistry.create({
-    runId: createRuntimeRunId(),
-    status: "queued",
-    metadata: {
-      kind: "recipe-run",
-      recipePath,
-      artifactsDirectory: configuredArtifactsDirectory,
-    },
-    replay: {
-      command: ["wp-codebox", "recipe-run", "--recipe", recipePath],
-      recipePath,
-    },
-  })
-  const artifactPointer = new RecipeArtifactPointerTracker(resolve(configuredArtifactsDirectory ?? "artifacts"), runRecord.runId, recipePath, new Date(startedAtMs).toISOString())
+  const context = await createRecipeRunContext(options)
+  const { recipePath, recipeDirectory, recipe, configuredArtifactsDirectory, runRegistry, artifactPointer, startedAtMs } = context
+  let { runRecord } = context
   await artifactPointer.update({ commandStatus: "queued" })
   const issues = await validateWorkspaceRecipe(recipe, recipePath)
   if (issues.length > 0) {
@@ -846,7 +133,7 @@ async function runRecipe(options: RecipeRunOptions, interruption?: RecipeInterru
   let artifacts: ArtifactBundle | undefined
   let startupDurationMs: number | undefined
   let cleanupEvidence: RunResourceCleanupEvidence | undefined
-  const phaseTracker = new RecipePhaseTracker()
+  const phaseTracker = new RecipePhaseTracker(serializeRecipeRunError, isRecipeRunTimeoutError)
   const awaitRecipe = <T>(operation: string, promiseOrFactory: Promise<T> | (() => Promise<T>), timeoutMs = remainingRecipeTimeoutMs(startedAtMs, options.timeoutMs)): Promise<T> => {
     return artifactPointer.update({ command: operation, commandStatus: "running", phases: phaseTracker.list() })
       .then(() => {
@@ -1317,18 +604,6 @@ function cleanupEvidenceFromRunRecord(runRecord: RuntimeRunRecord, durationMs: n
     attempts: cleanup.attempts,
     error: error ?? cleanup.error,
   }) as RunResourceCleanupEvidence
-}
-
-function recipeRunFailureStatus(error: unknown, interruption?: RecipeInterruptionController): RuntimeRunRecord["status"] {
-  if (interruption?.metadata) {
-    return "cancelled"
-  }
-
-  if (error instanceof RecipeRunTimeoutError) {
-    return "timed_out"
-  }
-
-  return "failed"
 }
 
 async function runResourceEvidence(options: RunResourceEvidenceOptions): Promise<Record<string, unknown>> {
